templates/python/ai-bot/src/requirements.txt
flask[async]
teams_ai-0.0.0-py3-none-any.whl
semantic_kernel==0.3.7.dev0